refactor(#3038): move turn bridge voice completion helpers
